# hawkular-inventory-rest-api/src/main/rest-doc/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "${project.version}",
    "title" : "Hawkular Inventory"
  },
  "basePath" : "/hawkular/inventory",
...
  "paths" : {
...
    "/entity/{path}/history" : {
      "get" : {
        "tags" : [ "Single Entity" ],
        "summary" : "Obtains the history of the entity.",
        "description" : "",
        "operationId" : "getHistory",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "path",
          "in" : "path",
          "required" : true,
          "type": "string",
          "description" : "See the documentation above for the format of the path."
        }, {
          "in" : "query",
          "name" : "from",
          "description" : "Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        }, {
          "in" : "query",
          "name" : "to",
          "description" : "Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "History returned.",
            "schema" : {
              "$ref" : "#/definitions/History"
            }
          },
          "404" : {
            "description" : "No entity found on given traversal URI."
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
...
    "/tenant" : {
      "get" : {
        "tags" : [ "Tenant Information" ],
        "summary" : "Retrieves the details of the current tenant.",
        "description" : "",
        "operationId" : "getTenant",
        "parameters" : [ {
          "in" : "query",
          "name" : "at",
          "description" : "Look at the tenant as it looked at the provided time. Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        }],
        "responses" : {
          "200" : {
            "description" : "Tenant obtained",
            "schema" : {
              "$ref" : "#/definitions/Tenant"
            }
          },
          "403" : {
            "description" : "Authorization error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          },
          "500" : {
            "description" : "Server error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Tenant Information" ],
        "summary" : "Updates the properties of the tenant",
        "description" : "",
        "operationId" : "putTenant",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TenantUpdate"
          }
        }, {
          "in" : "query",
          "name" : "at",
          "description" : "The time to mark the update as having occured at. Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        } ],
        "responses" : {
          "204" : {
            "description" : "Tenant updated"
          },
          "403" : {
            "description" : "Authorization error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          },
          "500" : {
            "description" : "Server error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/tenant/relationship" : {
      "post" : {
        "tags" : [ "Tenant Information" ],
        "summary" : "Creates new relationship(s) on a tenant",
        "description" : "As all POSTs, this can also accept an array of blueprints to create more than 1 relationship at a time.",
        "operationId" : "createRelationships",
        "parameters" : [ {
          "in" : "query",
          "name" : "at",
          "description" : "The time to mark the relationship as having been created at. Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        }],
        "responses" : {
          "201" : {
            "description" : "Relationship(s) created.",
            "schema" : {
              "$ref" : "#/definitions/Relationship"
            }
          },
          "400" : {
            "description" : "Data in wrong format"
          },
          "403" : {
            "description" : "Unauthorized access"
          },
          "404" : {
            "description" : "One of the related entities doesn't exist."
          },
          "500" : {
            "description" : "Server error."
          }
        }
      }
    },
    "/tenant/relationships.*" : {
      "get" : {
        "tags" : [ "Tenant Information" ],
        "summary" : "Retrieves tenant's relationships",
        "description" : "By default, all the outgoing relationships are returned. This can be modified using the same techniques as in the entity graph traversal.",
        "operationId" : "getRelationships",
        "parameters" : [ {
          "in" : "query",
          "name" : "at",
          "description" : "Look at the relationships as they were at the provided time. Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        }],
        "responses" : {
          "200" : {
            "description" : "Relationships returned",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Relationship"
              }
            }
          },
          "403" : {
            "description" : "Authorization error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          },
          "500" : {
            "description" : "Server error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/tenant/history" : {
      "get" : {
        "tags" : [ "Tenant information" ],
        "summary" : "Obtains the history of the updates to the tenant.",
        "description" : "",
        "operationId" : "getTenantHistory",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "path",
          "in" : "path",
          "required" : true,
          "type": "string",
          "description" : "See the documentation above for the format of the path."
        }, {
          "in" : "query",
          "name" : "from",
          "description" : "Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        }, {
          "in" : "query",
          "name" : "to",
          "description" : "Either a Unix epoch timestamp in milliseconds, or an ISO-8601 formatted date and time in UTC timezone.",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "History returned.",
            "schema" : {
              "$ref" : "#/definitions/History"
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "AbstractElement" : {
      "type" : "object",
      "properties" : {
        "path" : {
          "$ref" : "#/definitions/CanonicalPath"
        },
        "properties" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "id" : {
          "type" : "string"
        }
      },
      "required": ["path"]
    },
...
    "AbstractElementUpdate" : {
      "type" : "object",
      "properties" : {
        "properties" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "ApiError" : {
      "type" : "object",
      "properties" : {
        "errorMsg" : {
          "type" : "string",
          "description" : "Detailed error message of what happened",
          "readOnly" : true
        },
        "details" : {
          "type" : "object",
          "description" : "Optional details about the error beyond what's provided in the error message.",
          "readOnly" : true
        }
      },
      "description" : "If REST-call returns other than success, detailed error is returned."
    },
    "CanonicalPath" : {
      "type" : "string",
      "description" : "A canonical path is slash-separated list of path segments that uniquely identity an entity in the Hawkular Inventory graph. The path follows the \"contains\" relationships in inventory (which form a tree structure). Each segment specifies the type of the element on that position in the tree, followed by semicolon and the ID of the element. An example of a canonical path would be \"/t;tenant/f;my-feed/r;my-resource\". The type is one of 't' (tenant), 'e' (environment), 'rt' (resource type), 'mt' (metric type), 'ot' (operation type), 'mp' (metadata pack), 'r' (resource), 'm' (metric), 'd' (data) or 'rl' (relationship). Please consult Hawkular Inventory documentation for a more thorough discussion of the different types of entities and their places in the model."
    },
...
    "Entity" : {
      "allOf" : [ {
        "$ref" : "#/definitions/AbstractElement"
      }, {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        },
        "description" : "Defines the basic properties of all entity types in inventory"
      } ]
    },
...
    "EntityUpdate" : {
      "allOf" : [ {
        "$ref" : "#/definitions/AbstractElementUpdate"
      }, {
        "type" : "object",
        "properties" : {
          "name": {
            "type": "string"
          }
        }
      } ]
    },
...
    "Relationship" : {
      "allOf" : [{
        "$ref" : "#/definitions/AbstractElement"
      }, {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "source" : {
            "$ref" : "#/definitions/CanonicalPath"
          },
          "target" : {
            "$ref" : "#/definitions/CanonicalPath"
          }
        },
        "required": ["id", "source", "target"],
        "description" : "A relationship between two entities."
      }]
    },
...
    "Tenant" : {
      "allOf" : [ {
        "$ref" : "#/definitions/Entity"
      }, {
        "type" : "object",
        "description" : "The tenants partition the data in the inventory graph. No relationships between entities from 2 different tenants can exist."
      } ]
    },
    "TenantUpdate" : {
      "allOf": [{
        "$ref": "#/definitions/EntityUpdate"
      }]
    },
...
    "History" : {
      "type" : "array",
      "items" : {
        "type" : "Change"
      }
    },
...
  }
}